Backend/churn-backend-integrated/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.20
SQLAlchemy>=1.4
asyncpg>=0.27
alembic>=1.10
pydantic>=1.10
python-dotenv>=1.0
psycopg2-binary>=2.9  # for migrations/tools (async driver used at runtime is asyncpg)
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0
google-generativeai>=0.3.0
reportlab>=4.0.0
APScheduler>=3.10.0
python-dateutil>=2.8.0
